organise AI functions into strategies
